services/graph-bot/requirements.txt
fastapi
uvicorn
httpx
pydantic
chromadb
langgraph
google-genai
langchain
langchain-neo4j
langchain-core
langchain-google-genai
langsmith
neo4j